- do not try to backup nonexisting /boot/initrd
Changed files:
    kernel.spec -> 1.201

+++ kernel.spec
@@ -679,7 +679,7 @@ ln -sf vmlinuz-%{version}-%{release} /boot/vmlinuz
 ln -sf System.map-%{version}-%{release} /boot/System.map
 
 geninitrd /boot/initrd-%{version}-%{release}.gz %{version}-%{release}
-mv -f /boot/initrd /boot/initrd.old 2> /dev/null > /dev/null
+test ! -f /boot/initrd || mv -f /boot/initrd /boot/initrd.old 2> /dev/null > /dev/null
 ln -sf initrd-%{version}-%{release}.gz /boot/initrd
 
 if [ -x /sbin/lilo -a -f /etc/lilo.conf ]; then
@@ -696,7 +696,7 @@ ln -sf vmlinuz-%{version}-%{release}smp /boot/vmlinuz
 ln -sf System.map-%{version}-%{release}smp /boot/System.map
 
 geninitrd /boot/initrd-%{version}-%{release}smp.gz %{version}-%{release}smp
-mv -f /boot/initrd /boot/initrd.old 2> /deb/null > /dev/null
+test ! -f /boot/initrd || mv -f /boot/initrd /boot/initrd.old 2> /deb/null > /dev/null
 ln -sf initrd-%{version}-%{release}smp.gz /boot/initrd
 
 if [ -x /sbin/lilo -a -f /etc/lilo.conf ]; then
